# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LeptonPad API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  walletAddress: string | null;
  selectedCategories: string[];
  onboardingComplete: boolean;
  verified: boolean;
  isAdmin: boolean;
  createdAt: string;
}

export interface UserUpdate {
  selectedCategories?: string[];
  onboardingComplete?: boolean;
  name?: string;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  description: string;
  icon: string;
}

export type ContentType = typeof ContentType[keyof typeof ContentType];


export const ContentType = {
  article: 'article',
  audio: 'audio',
  video: 'video',
} as const;

export interface Content {
  id: number;
  title: string;
  type: ContentType;
  categorySlug: string;
  categoryName: string;
  price: number;
  creatorId: string;
  creatorName: string;
  /** @nullable */
  creatorImageUrl: string | null;
  creatorVerified: boolean;
  /** @nullable */
  previewText: string | null;
  /** @nullable */
  coverImageUrl: string | null;
  /** @nullable */
  audioUrl: string | null;
  /** @nullable */
  videoUrl: string | null;
  viewCount: number;
  purchaseCount: number;
  published: boolean;
  createdAt: string;
}

export type ContentDetailType = typeof ContentDetailType[keyof typeof ContentDetailType];


export const ContentDetailType = {
  article: 'article',
  audio: 'audio',
  video: 'video',
} as const;

export interface ContentDetail {
  id: number;
  title: string;
  type: ContentDetailType;
  categorySlug: string;
  categoryName: string;
  price: number;
  creatorId: string;
  creatorName: string;
  /** @nullable */
  creatorImageUrl: string | null;
  /** @nullable */
  body: string | null;
  /** @nullable */
  previewText: string | null;
  /** @nullable */
  coverImageUrl: string | null;
  /** @nullable */
  audioUrl: string | null;
  /** @nullable */
  videoUrl: string | null;
  viewCount: number;
  purchaseCount: number;
  /** @nullable */
  readingTimeMinutes: number | null;
  published: boolean;
  createdAt: string;
  hasAccess: boolean;
  creatorVerified: boolean;
}

export type ContentInputType = typeof ContentInputType[keyof typeof ContentInputType];


export const ContentInputType = {
  article: 'article',
  audio: 'audio',
  video: 'video',
} as const;

export interface ContentInput {
  /** @minLength 1 */
  title: string;
  type: ContentInputType;
  categorySlug: string;
  body?: string;
  previewText?: string;
  coverImageUrl: string;
  audioUrl?: string;
  videoUrl?: string;
  /** @minimum 0 */
  price: number;
  published?: boolean;
}

export interface ContentUpdate {
  title?: string;
  body?: string;
  previewText?: string;
  coverImageUrl?: string;
  audioUrl?: string;
  videoUrl?: string;
  /** @minimum 0 */
  price?: number;
  published?: boolean;
}

export interface ContentListResponse {
  items: Content[];
  total: number;
  offset: number;
  limit: number;
}

export interface UnlockInput {
  contentId: number;
  /** @nullable */
  txHash?: string | null;
}

export interface UnlockResult {
  success: boolean;
  paymentId: number;
  contentId: number;
  amountPaid: number;
  creatorReceives: number;
  settledAt: string;
}

export interface AccessStatus {
  hasAccess: boolean;
  /** @nullable */
  paymentId: number | null;
}

export interface Purchase {
  id: number;
  contentId: number;
  contentTitle: string;
  contentType: string;
  creatorName: string;
  amountPaid: number;
  paidAt: string;
}

export interface EarningsSummary {
  totalEarnedAllTime: number;
  thisWeekEarnings: number;
  totalPurchases: number;
  totalViews: number;
  conversionRate: number;
}

export interface ContentEarnings {
  contentId: number;
  title: string;
  type: string;
  price: number;
  views: number;
  purchases: number;
  conversionRate: number;
  totalEarned: number;
  hasSuggestion: boolean;
  publishedAt: string;
}

export interface PaymentEvent {
  id: number;
  contentId: number;
  contentTitle: string;
  readerName: string;
  amount: number;
  creatorAmount: number;
  paidAt: string;
}

export type AiSuggestionAction = typeof AiSuggestionAction[keyof typeof AiSuggestionAction];


export const AiSuggestionAction = {
  raise: 'raise',
  lower: 'lower',
  keep: 'keep',
} as const;

export type AiSuggestionStatus = typeof AiSuggestionStatus[keyof typeof AiSuggestionStatus];


export const AiSuggestionStatus = {
  pending: 'pending',
  applied: 'applied',
  dismissed: 'dismissed',
} as const;

export interface AiSuggestion {
  id: number;
  contentId: number;
  contentTitle: string;
  currentPrice: number;
  suggestedPrice: number;
  action: AiSuggestionAction;
  reasoning: string;
  conversionRate: number;
  status: AiSuggestionStatus;
  createdAt: string;
}

export interface VerifyInput {
  verified: boolean;
}

export interface PlatformStats {
  totalUsdcPaid: number;
  totalCreators: number;
  totalContent: number;
  totalPayments: number;
  since: string;
}

export type ListContentParams = {
/**
 * Comma-separated category slugs to filter by
 */
categories?: string;
/**
 * Filter by content type
 */
type?: ListContentType;
/**
 * Filter by creator Clerk user ID
 */
creatorId?: string;
/**
 * Search titles, previews, and creator names
 */
q?: string;
limit?: number;
offset?: number;
};

export type ListContentType = typeof ListContentType[keyof typeof ListContentType];


export const ListContentType = {
  article: 'article',
  audio: 'audio',
  video: 'video',
} as const;

export type GetRecentPaymentsParams = {
limit?: number;
};