# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeptonPad API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{clerkId}/verify:
    post:
      operationId: verifyCreator
      tags: [users]
      summary: Grant or revoke verification badge (admin only)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (not admin)
        "404":
          description: User not found
...
  /content/{id}:
    get:
      operationId: getContent
      tags: [content]
      summary: Get single content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "404":
          description: Not found
    put:
      operationId: updateContent
      tags: [content]
      summary: Update content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdate"
      responses:
        "200":
          description: Updated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        imageUrl:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
        selectedCategories:
          type: array
          items:
            type: string
        onboardingComplete:
          type: boolean
        verified:
          type: boolean
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - name
        - email
        - imageUrl
        - walletAddress
        - selectedCategories
        - onboardingComplete
        - verified
        - isAdmin
        - createdAt
...
    Content:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [article, audio, video]
        categorySlug:
          type: string
        categoryName:
          type: string
        price:
          type: number
        creatorId:
          type: string
        creatorName:
          type: string
        creatorImageUrl:
          type: ["string", "null"]
        creatorVerified:
          type: boolean
        previewText:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        viewCount:
          type: integer
        purchaseCount:
          type: integer
        published:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - type
        - categorySlug
        - categoryName
        - price
        - creatorId
        - creatorName
        - creatorImageUrl
        - creatorVerified
        - previewText
        - coverImageUrl
        - audioUrl
        - videoUrl
        - viewCount
        - purchaseCount
        - published
        - createdAt
    ContentDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [article, audio, video]
        categorySlug:
          type: string
        categoryName:
          type: string
        price:
          type: number
        creatorId:
          type: string
        creatorName:
          type: string
        creatorImageUrl:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        previewText:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        viewCount:
          type: integer
        purchaseCount:
          type: integer
        readingTimeMinutes:
          type: ["integer", "null"]
        published:
          type: boolean
        createdAt:
          type: string
        hasAccess:
          type: boolean
        creatorVerified:
          type: boolean
      required:
        - id
        - title
        - type
        - categorySlug
        - categoryName
        - price
        - creatorId
        - creatorName
        - creatorImageUrl
        - creatorVerified
        - body
        - previewText
        - coverImageUrl
        - audioUrl
        - videoUrl
        - viewCount
        - purchaseCount
        - readingTimeMinutes
        - published
        - createdAt
        - hasAccess
...
    ContentUpdate:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        previewText:
          type: string
        coverImageUrl:
          type: string
          description: Uploaded file path or absolute URL
        audioUrl:
          type: string
        videoUrl:
          type: string
        price:
          type: number
          minimum: 0
        published:
          type: boolean
...
    VerifyInput:
      type: object
      properties:
        verified:
          type: boolean
      required:
        - verified
...